+++ ng2-metadata.ts
@@ -1,0 +1,4 @@
+export * from './dist/models/page-title-positioning';
+export * from './dist/models/metadata-settings';
+export * from './dist/metadata.service';
+export * from './dist/metadata.module';