+++ src/index.ts
@@ -1,4 +1,0 @@
-export { PageTitlePositioning } from './models/page-title-positioning';
-export { MetadataSettings } from './models/metadata-settings';
-export { MetadataService } from './metadata.service';
-export { MetadataModule } from './metadata.module';